backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
flask_sqlalchemy
flask_jwt_extended
flask_migrate